requirements.txt
streamlit
pytesseract
numpy
pdf2image
Pillow